6. Deployment/requirements_pipfreeze.txt
absl-py==1.0.0
anyio==3.5.0
asgiref==3.5.0
async-generator==1.10
attrs==21.4.0
backcall==0.2.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
click==8.0.3
colorama==0.4.4
convertdate==2.4.0
cryptography==36.0.1
cycler==0.11.0
decorator==5.1.1
Deprecated==1.2.13
fastapi==0.73.0
fonttools==4.29.1
gcloud==0.18.3
google-auth==2.6.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.54.0
grpcio==1.43.0
h11==0.13.0
hijri-converter==2.2.2
holidays==0.12
httplib2==0.20.2
idna==3.3
importlib-metadata==4.10.1
ipython==7.31.1
jedi==0.18.1
joblib==1.1.0
jwcrypto==1.0
kiwisolver==1.3.2
korean-lunar-calendar==0.2.1
lightgbm==3.3.2
Markdown==3.3.6
matplotlib==3.5.1
matplotlib-inline==0.1.3
nfoursid==0.1.0
numpy==1.21.5
oauth2client==4.1.3
oauthlib==3.2.0
outcome==1.1.0
packaging==21.3
pandas==1.3.5
parso==0.8.3
patsy==0.5.2
pickleshare==0.7.5
Pillow==9.0.0
pip==21.2.4
prompt-toolkit==3.0.26
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.14.0
pydantic==1.9.0
Pygments==2.11.2
PyMeeus==0.5.11
pyOpenSSL==22.0.0
pyparsing==3.0.7
Pyrebase4==4.5.0
python-dateutil==2.8.2
python-jwt==3.3.2
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rsa==4.8
scikit-learn==1.0.2
scipy==1.7.3
selenium==4.1.0
setuptools==58.0.4
six==1.16.0
sklearn==0.0
sniffio==1.2.0
sortedcontainers==2.4.0
starlette==0.17.1
statsmodels==0.13.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
torch==1.10.2
tqdm==4.62.3
traitlets==5.1.1
trio==0.19.0
trio-websocket==0.9.2
typing_extensions==4.0.1
u8darts==0.16.1
urllib3==1.26.8
uvicorn==0.17.1
wcwidth==0.2.5
Werkzeug==2.0.2
wheel==0.37.1
wincertstore==0.2
wrapt==1.13.3
wsproto==1.0.0
xarray==0.20.2
zipp==3.7.0
